my-dashboard/requirements.txt
streamlit
pandas
numpy
pymysql
feedparser
beautifulsoup4
openai
requests
